Change scheduler::Message::{Client,Worker}Connected to take a sender.
Before, the scheduler messages didn't take a sender at all, and the
Scheduler passed the ID to the SchedulerDeps when it wanted to send a
message. This resulted in the adapter having to maintain its own hash
table and translate IDs to senders. This change removes all of that
duplication and also reduces the amount of code in src/broker.rs, which
isn't unit tested.

To accomplish this, the two receiver types had to be added to
SchedulerDeps, which resulted in Scheduler, Worker, and Message all
having to become generic types. The Scheduler treats these receiver
types as completely opaque: it just stores them in its hash table and
passes them to the SchedulerDeps to send messages.