Avoid using one character generics
Its okay to use them when their scope is a single line, but for
functions and especially impl blocks it can be really confusing.

This is something many other people are bad at, but I think giving them
names its just strictly better, and definitely not unheard of.

I like to use the 'T' suffix with them so you still know they are
template parameters and don't confuse them with concrete types.