Provide default implementations for CacheDeps trait.
Also change Rng and rng() to just rand_u64 in the crate.

Also change read_dir to return a Box<dyn Iterator<Item = PathBuf>>
instead of a static type. This simplifies things a lot and
we really don't care about the performance of method: it's
only used at startup time, and the performance impact will
be negligible.