Move some code down into the spawned task in broker's listener_main.
When we implement it, proto::Hello::{Client,Worker}Artifact
will be handled on *threads* using the standard library, not
on *tasks* using Tokio. The reason for this is that we'll be
able to more easily interact with tar and gzip libraries,
and also so we can use sendfile(2) or splice(2) under the
hood.

This refactor delays the splitting of the (tokio) socket
into two and the addition of the (tokio) io::BufReader until
we know we are speaking Hello::{Client,Worker}.

In the future, when handling Hello::{Client,Worker}Artifact,
we will convert the tokio socket into an std socket, make it
blocking, and continue on.