Remove PathBuf from GetArtifact::Get.
Originally I thought it would make sense for the Cache to give out a Path for
the file to go into. Upon further reflection, this would require us keeping
that Path somewhere and somehow associating it with incoming artifact pushes
from clients.

It seems to make more sense to just provide a method on the Cache to get the
directory that artifacts will go into, and then use an off-the-shelf tmpfile
crate.