Add clone_into_pid_and_user_namespace.
Now the worker is two processes. The parent process waits for the child process
and mimics the latter's termination. The child sets up a parent death signal on
the parent so it will be killed if the parent terminates.

The child also runs in its own user and pid namespaces. Because it's pid 1 in
its own pid namespace, all of the other processes in the pid namespace will be
immediately killed with SIGKILL (9) if the child terminates. This guarantees
that we won't leak processes when shutting down the worker.

We need the user namespace so we can create the pid namespace. The worker is
uid 0 and gid 0 in the user namespace.

The worker now also maps stdin to /dev/null (effectively), so that it won't
have to do so when starting jobs.