Create a new meticulous-worker-child crate.
This will be a no_std crate used in the child process of the worker.

It'll be no_std because after we clone we can't use the allocator or anything
else where a thread in the parent may have held a lock at clone time.

This initial version of the crate isn't no_std, but we'll work our way there.