Change the way executor tests work to always use layers.
Before, the tests themselves called into the executor with no layers, which
meant the code they passed into the executor was run in the enclosing
namespace. For cargo test, this was the host machine's environment. For cargo
metest, this was specified by the executor-test-deps.tar file.

With this change, the tests always give the executor the executor-test-deps.tar
file as a layer, so the code is always running in its own container. This means
we need to embed the tar file into the tests themselve. But it also means that
we don't need a very big environment for the test itself to run in (just /tmp
and /proc).