Start changing worker-child to take a syscall script.
We're going to create a "script" of syscalls in the parent before the clone,
and then hand it to the client to just run. This should be easier on the child
than trying to execute a bunch of logic without allocating. It will also give
us a testing interface.

This first commit just starts at the bottom and moves over the unmount.