Start switching over to a child mechanism where we only send it syscalls to execute.
With this new design, the child will just get a "script" of syscalls to run.
We'll build the script in the parent before the clone. This will minimize the
mount of work we need to do in the child in the weird no-alloc environment. It
will also reduce the amount of data we have to pass to the child.