Move JobDescription out of main.rs into the library.
I'm not 100% happy with this. I think we'll probably want to rename either the
binary or the library. This is library code that is only used by this binary.
But that's a topic for another day.

The next step is to add better parsing of this structure. It used to be called
JobDescription, and now it's just called Job.