Change the way cargo-metest deals with being called from cargo.
Before this change, we told clap that we were a binary called `cargo` that
expected exactly one subcommand called `metest`. There were two downsides to
this approach.

First, it required that the binary always be called `cargo-metest`. If we
decided to install it as `cargo-mt` or something else, this would break. I'm
currently testing it by having a symlink for `cargo-mt` that points to
`cargo-metest`. In the real world, we're probably tell our users to use cargo
aliases, but it's still nice to know the binary is flexible.

Second, it made it weird to run the binary outside of cargo. If you ran
`cargo-metest` by itself, you would get an error. You had to run `cargo-metest
metest`, which is weird.

The new approach is to make a copy of env::args(), possibly modify it, and then
pass the copy off to clap for parsing. If the second argument matches the tail
end of the first, with `cargo-` prepended, then we just remove the second
argument.