Keep the ExecutingJob around even after it's canceled.
We get rid of the canceled HashMap. In its stead, we add an ExecutingJobState
enum. ExecutingJobState::Ok represents what was formerly the only state for an
Executing job: it's running and hasn't been canceled. The new
ExecutingJobState::Canceled represents the job after it's been canceled, but
before the status, stdout, and stderr have been collected.

This does a few things. First, it allows us to send the stdout and stderr back
in response to a canceled job. We may want to do this eventually for
bookkeeping that could be used for billing or reporting.

Second, it makes it easy to add a TimedOut state where we will want to return
stdout and stderr.

Third, it does cause one behavioral change. A new job will not start until the
old job has completely finished. It's not clear if this is a good thing or a
bad thing, but it probably doesn't matter much one way or the other.