Work around strange exec ETXTBSY issue
If you try to exec a file which is open for writing you get ETXTBSY. We
had the file open for writing just immediately before, but we closed it.
I have no idea why Linux thinks this file is still open for writing, so
we just retry until it works.