Rework how command-line options, environment variables, and config settings are parsed in maelstrom-worker.
This approach forgoes figment and replaces it with something home-grown. That
home-grown "something" is in the new maelstrom-config crate.

The idea is to have a single struct where all config is put, and to enable some
nice features like adding the environment variables and config-file keys to
help messages. There's still a lot to do, but this is a start.

Eventually, we're going to want to generate a lot of this code with derive proc
macros.