Flush locked tags file after writing it
This seems to fix a rare test bug where the file contents are not
observable directly after get_container_image function call.

AFAIK the file should have been closed at that point and the contents
written, but maybe it gets delayed somehow in the compiler generated
async code or the kernel.