Start porting cargo-maelstrom over to the new maelstrom_config library.
This first changeset just picks off the easy arguments. There is still
functionality missing in maelstrom_config that will be added in the upcoming
changsets.

I created a new_config2 function that gives a hook for adding additional
command line arguments as well as getting the ArgMatches back in the end.