Change clone_into_pid_and_user_namespace to clone into a grandchild intead of a child.
The motivation here is that having the worker also be pid 1 in the new pid
namespace causes problems that result in an ugly design for the worker. By
making the worker not be pid 1, we can simplify the worker design. In
particular, the worker can call waitpid or equivalent safely and not worry
about the accumulation of zombie descendants.