Use new mechanism for reaping children processes in the worker.
In the old world, we had a separate reaper thread that would just loop, calling
wait. It would then send the reaped children's pids to the dispatcher. In the
new world, we're no longer the init process for the pid namespace, so we don't
have to worry about reaping zombies. Now, each job just has its child process
reaped with waitpid (actually waitid with a pidfd).