Don't check if the client socket is closed in such a tight loop
This loop was burning a ton of CPU needlessly

Also, when we get an error from ready, just exit the future instead of
continue to wait. We can only get an error there when the runtime is
shutting down, so we should just exit.

5 seconds seems like a good amount of time, we only really need to do
this at all so that the client-process can exit when the client has gone
away