Introduce and use a new StreamWrapper type in maelstrom-client-process.
Tonic's interface is a little bit inconvenient for our use case. They want to
take a stream of streams, and a possible shutdown future. They will tear
everything down if the stream of streams ends, or if the shutdown future
returns.

What we want is it to serve until one single stream is closed by the remote
side. We achieve this by wrapping our stream with something that detects the
EOF and triggers an event. We can then pass that event as the shutdown future.