When we can't get pipe memory we need, fallback to using slower IO
I got really tired of the worker crashing due to exhausting pipe memory.
This can be tricky because if we're not careful it will be falling back
to the slower IO without us knowing. I made the worker log at error
level if this happens so we can hopefully get some insight