Change async_socket_writer and async_socket_reader to return a Result<()>.
This allows us to distinguish between two different cases: when we have an
actual error with the socket, and when the channel we're receiving from or
sending to has its other side closed. In the latter case we return Ok(()), but
in the former case we return the socket error.

The idea is that channels getting closed isn't really interesting. It means
there was an error with a different task, and it's exited as a result.