Change channel_reader to return a Result<()>.
This is always Ok(()), but it lets us use it in the same JoinSet as things that
fail. We could change channel_reader to take a processor closure that returns a
Result, and return Err if the processor does. However, we don't have any
processors that want to return errors yet.