Pass the container_image_depot_cache_dir into the client process.
We very quickly become root in our own namespace in the client process, so we
can't use xdg to get the result when we need it. In theory, we could get the
value before we go into our own namespace, but it seems cleaner to just have
the start message contain in.