Fix a race inside cargo-maelstrom
We weren't properly waiting for job's callback to be called ever since
the RPC boundary was added.

The race was avoided by waiting for the progress bars to complete, but
when a progress bar isn't being displayed it exists.

The new integration test caught this, since I wasn't using the progress
bar based output.