Remove wait_for_outstanding_jobs RPC
This RPC is error-prone because it is easy to think it will wait for all
your job RPCs to have completed, but it isn't able to do that.

Furthermore, if you have threads waiting for jobs to complete, this call
can't possibly wait for those to finish.

It is better to encourage some kind of client-side synchronization
instead.