Change VecDeque in worker's Dispatcher to be a BinaryHeap.
Jobs with the longest estimated duration will be chosen first. Jobs without any
estimated duration will be chosen before those with an estimated duration.

We're doing this as part of using LPT (Longest Processing Time first)
scheduling, which should give us better schedules. We'll also make a similar
change in the broker.

There are a few reasons to make the change in both the broker and the worker.

First, as it currently stands, the client process just routes all local job
directly to the worker. So, the local worker has to act like its own broker in
this regard.

Second, we oversubscribe the workers so that they will have work to pick up
while traffic transits the network. When the broker gets a new job that has a
long estimated duration, and then sends it to a worker, we want that job to go
to the head of the queue.