Plumb JobNetwork into maelstrom_base::JobSpec.
The JobNetwork::Local can never be sent to the broker. I was originally
planning on enforcing this by not even having it available in
maelstrom_base::JobSpec. However, that's not really going to work out since we
need it in the worker (at least the executor).

In theory, I guess we could create a JobSpec for on the wire from the client to
the broker, and then have a different one that the worker uses (and yet another
different one for the executor). But that seems like overkill for now.

Instead, I think I'm just going to validate on the broker that it never
receives a JobSpec that should only be executed locally. We could also add
something similar to the worker when it reads JobSpecs off the wire.