container: Avoid contending with ourselves for file-locks
Contending on a file-lock uses a file-descriptor, and we only have so
many of those.

When we have multiple tasks all trying to download a container image,
they all end up contenting on the file-locks and that can hit the limit
for open files.

Fix this by adding a mutex for in-process contending on filling the
cache.

fix: #291