Move image tag parsing to inside maelstrom-container
This removes the tag field from the various ImageSpecs, instead the
image name will just include the tag and in a future change any of the
other information.

Doing this parsing inside maelstrom-container is very convenient since
now the various front-ends to the client don't have to worry about
parsing valid image names.

This means that some validation is now happening later rather than
earlier, but I think the convenience of it outweighs this concern. You
have to try pretty hard to create an invalid image name.