Implement maelstrom-run --one.
You can run `maelstrom-run --one` with no extra arguments, in which case the
program and arguments are taken from the job spec. You can also run
`maelstrom-run --one program argument` in which case program and arguments from
the job spec are replaced with those provided on the command-line.

Currently, you still have to have a program in the job specification, even if
you're going to ignore it, since the parsing of the job specification requires
the program.

Fixes #302.