Change expect() to be more resillient to timing issues.
The new technique is to look at the output we're gotten so far and to check if
it's a prefix of the expected output. If it is a prefix, but not the whole
string, then we wait for a minute for more output. If it's a prefix, but
doesn't match, then we fail immediately.

The only downside of this approach is that if the actual output is a prefix of
the expected output, we'll wait for up to a minute before we fail.