Add feature to search PATH when looking for program to execute.
This is modeled on execvp, but we have our own implementation. There are a few
reasons for this.

First, execvp may do allocations or other unsafe things, so we can't call it
from the child process.

Second, to get execvp to function correctly, we'd have to set the global C
`environ` variable, which is just gross.

Third, execvp falls back on using a default path if you don't provide one, and
that default path is system-dependent. We want to have our own, well-specified,
implementation. In our particular case, this means that we only search the path
given in PATH.

Fixes #300.