Change how default for include_shared_libraries is determined.
If you set it to true or false explicitly, that's what it is. But if you don't
set it, we have some logic to determine what it should be. Before, the logic
was that if you had any layers included, then it should be false. That may have
made sense when we only had tar layers, but with all of the layer types we have
now, with stubs and symlinks in particular, it doesn't make sense anymore.

The new logic is that it will be true, unless you base container off of an
image. The idea is that the image should have any shared libraries you want,
and you're probably relying on that.

This may all become moot if/when we do specific layers for shared library
dependencies.

Fixes #232.