fix: #328: cargo-maelstrom Delay yielding cargo integration tests
We delay yielding them until all the binaries for their package are
finished building. This is because the integration tests are allowed to
use these binaries.

This should hopefully fix the issues we've seen where cargo-maelstrom is
trying to upload some package binary and it has been modified in the
middle of the process and this causes an error.