Rename "loop" to "repeat".
Loop doesn't seem quite right here. We're asking the test runner to repeat the
test a certain number of times. While we could ask it to "loop" the test or
tests, it doesn't seem to quite fit. To me, looping would almost suggest
running all of the tests to completion, then starting again, etc.