Fix small visual bug with the multi-gauge TUI element
This detection of being inside the gauge label was off by one on the
right side. I noticed this previously but since the original gauge code
I based this off of did this, I assumed there was some reason.

After seeing this bug happen though, I think it was just a mistake.