Fix issue with the fancy UI and the summary crahsing in ratatui
There is as far as I can tell a bug in Terminal::insert_before where if
the terminal area you are inserting into is too big, ratatui just
crashes. I ran into this when trying to print a summary that was too
large. I attempt to fix this here by printing each individual line of
the summary one at a time.