Re-arrange cache dir a little bit for worker's cache.
Now there are two top-level directories: removing/ and sha256/. Within sha256/
there is now one directory for each kind.

This removes any sort of conflict that could arise from having a kind with a
serialization of "removing" or any other thing we want to put in the top-level
cache directory.