Reimplement TestFs2::lookup to use a new function called lookup_index_path.
Because of Rust's safety guarantees, especially regarding references, we
weren't going to be able to make the old approach work for functions that
modified the file system. With this new approach, we can first resolve things
into a vector of indexes, which we can then use with mutable references to
actually make changes.