Remove unnecessary test implementation of Fs.
In the dispatcher tests, we don't really use any of the methods on the Fs
trait. Instead, we just need the TempFile and TempDirectory types so we can
pass them in the GotArtifactSuccess message.

There's no point having a whole test implementation of Fs that is just
unimplemented!. Instead, just add a few conveniences on the existing
test::TempFile and test::TempDirectory types and use them.