Remove read_manifest method from broker's cache.
This is part of the work of unifying the broker's and the worker's caches.
We're going to add features to the worker's cache until it is up to par with
the broker's cache, then switch the broker over.

The broker's cache really shouldn' have had this method. Instead, we just made
the cache_path method public and moved the reading of the file outside of the
cache code.

For the scheduler, this meant adding a new method on the deps to read a
manifest from a path.