Add the notion of a GetStrategy to the worker's cache.
This is one of the requirements that the broker has on its cache that the
worker doesn't.

The idea here is that, for the broker, it wants to request that an artifact be
fulfilled from each client. This is for redundancy's sake: we don't want one
client to get blocked behind a different, slow or misbehaving client.

To prepare the cache, we add the notion of a Getter and a GetStrategy. The
GetStrategy provides the Getter type and a function to map JobIds to Getters.

The Getter represents where an artifact is coming from. The cache now keeps a
set of all getters that it's currently waiting on for an entry. When a new get
comes in for an entry that already has a get outstanding on it, the cache won't
always return Wait anymore. Instead, it sees if the new job represents a new
Getter, and if so, it returns Get.

For the worker, the Getter is just (). For the broker, it's going to be
ClientId.