Change how logging of messages is done to be consistent.
Before, logging of messages sent and received was done in an ad-hoc manner. The
log messages would sometimes indicate which connection the message was sent to
or received from.

With this change, the maelstrom_util::net functions now all automatically log
messages sent and received, in a uniform way. To indicate which connection a
message is associated with, the caller now needs to add KV pairs to the logger.
The broker now does this by giving artifact fetcher and artifact pusher threads
their own pseudo IDs (afid and apid, respectively).

The alternate Debug format is used for logging messages, which leads to very
readable logs. However, the old alternate format for Sha256Digest was way too
long. This patch changes it to be identical to the non-alternate format.