Remove use of select! in the worker.
Now that the dispatcher fuses when shut down, we can just have the broker
sender task write straight to the dispatcher channel instead of having its own
channel and then having to use select.

This is one patch in a series that aims to unify the local worker with the "real" worker.