Broker reads manifests in an async way
It kicks off a job to read a manifest which sends the read digests back
to the scheduler via its message queue.

This is better than previously because it doesn't do IO in the
scheduler.

Local file IO isn't "blocking" in the traditional sense, but it can take
a while sometimes. Also, if we want our manifests to live on some
network backed cache, we will have to do real blocking in this path
anyway.