Remove DepsT template parameter from Message
Unfortunately this trick with using associated types when declaring
types is annoying in some ways.

First it can easily create weird bounds on the deps itself
like it being 'static, Send, or Sync.

Second it can cause unnecessary templating by requiring users to choose
certain types when it isn't needed. In particular here I want to avoid
having to template the message on the ArtifactStream type since it isn't
used in a message.