Add some robustness to CollapsedJobSpec::new implementation.
First, add checking for parent loops, and test for it.

Second, add a test for an unknown parent.

Third, short circuit parent evaluation based on what fields could possibly be
included from a parent. If we ever get to a point where we can prove that we're
not going to need any more parents, stop. This is similar to what we already do
with images.