Change how the client protocol specifies root-overlay.
The base JobSpec has a single field called root_overlay. This can be one of
None, Tmp, or Local. The latter contains two fields to indicate what directory
the overlayfs should use.

Before this change, the RootOverlay type, which is what the root_overlay field
is, was propagated all the way up to the client's ContainerSpec. The problem
was that this didn't really make much sense from the point of view of how
clients wanted to specify this field. From a client's point of view, they
really just want to have an `enable_writable_file_system` flag. The Local value
is really a more hidden thing that will hopefully, eventually, go away or at
least be specified differently.

To that end, this patch changes the client's JobSpec to have a separate field
called capture_file_system_changes. This is on the JobSpec, not on the
ContainerSpec. The idea is that this will only be used by maelstrom-pytest to
generate the container used to run tests in. When command layers are working,
this will go away completely.

This patch also changes the root_overlay field on the client's ContainerSpec to
be an enable_writable_file_system bool field. This is how the clients actually
think about things, so it matches well.