Replace BrokerToClient::ArtifactTransferredResponse with BrokerToClient::GeneralError.
This new message lets the broker tell the client about an error, with the
assumption that the client will shut down when it receives the message. In
theory, we could log the message at a high level as well, but this type of
error will likely cause a whole slew of individual jobs to fail.