Introduce ClientProcessFactory trait.
This is the beginning of a set of patches that will allow us to have multiple
Client instances in the same process. We will use this feature with the
`--watch` flag to create a whole new Client each time through instead of trying
to reuse a single Client.

With this change, instead of creating a background client process early in a
program's lifetime --- before it becomes multi-threaded --- we will instead
create a factory. That factory will then be able to be used at the point of
creation of the Client to create a new background process.

With this first patch, we don't actually do anything differently for the fork
path. Instead, one is only allowed to create one client process with the
factory. This will change in an upcoming patch.