Rewrite Watcher to not use scoped threads.
We don't really need to create a thread in the first place.

The main reason for this change is that I want the Watcher to be able to
out-live the Client, and that won't work with scoped threads.

There is another change here which is that the Watcher no longer pauses waiting
for changes. Instead, that has to be added externally.