Remove close() from BrokerSender.
It's unclear why this was even there. It might be because the broker connection
closing might initiate a shutdown, but that line of reasoning doesn't really
make sense. We can get into situations where the broker doesn't read messages,
and instead sends us jobs, and we can end up running out of memory. The point
is, we will have to deal with flow control from the broker at some point.