There is no point in printing out the PID, since it's always 2.
This is because we're in our own PID namespace. If we really wanted this info,
we'd have to pass it down from the grandparent process to the parent process,
which would be a pain. Another alternative would be to give the grandparent
PID, which would similarly require passing it down, but it would be easier to
come by. Anyway, it seems like ps or pstree works just fine.